+++ packages/core/src/CompositionManager.tsx
@@ -1,6 +1,5 @@
 import type {ComponentType, LazyExoticComponent} from 'react';
 import React, {
-	createContext,
 	useCallback,
 	useImperativeHandle,
 	useLayoutEffect,
@@ -11,6 +10,11 @@ import React, {
 import type {AnyZodObject} from 'zod';
 import {SharedAudioContextProvider} from './audio/shared-audio-tags.js';
 import type {CalculateMetadataFunction} from './Composition.js';
+import type {
+	BaseMetadata,
+	CompositionManagerContext,
+} from './CompositionManagerContext.js';
+import {CompositionManager} from './CompositionManagerContext.js';
 import type {TFolder} from './Folder.js';
 import type {
 	PropsIfHasProps,
@@ -136,54 +140,6 @@ export type TAsset = {
 	playbackRate: number;
 	allowAmplificationDuringRender: boolean;
 };
-
-type BaseMetadata = Pick<
-	AnyCompMetadata,
-	'durationInFrames' | 'fps' | 'defaultProps' | 'height' | 'width'
->;
-
-export type CompositionManagerContext = {
-	compositions: AnyComposition[];
-	registerComposition: <
-		Schema extends AnyZodObject,
-		Props extends Record<string, unknown> | undefined
-	>(
-		comp: TComposition<Schema, Props>
-	) => void;
-	unregisterComposition: (name: string) => void;
-	registerFolder: (name: string, parent: string | null) => void;
-	unregisterFolder: (name: string, parent: string | null) => void;
-	currentComposition: string | null;
-	setCurrentComposition: (curr: string) => void;
-	setCurrentCompositionMetadata: (metadata: BaseMetadata) => void;
-	currentCompositionMetadata: BaseMetadata | null;
-	registerSequence: (seq: TSequence) => void;
-	unregisterSequence: (id: string) => void;
-	registerAsset: (asset: TAsset) => void;
-	unregisterAsset: (id: string) => void;
-	sequences: TSequence[];
-	assets: TAsset[];
-	folders: TFolder[];
-};
-
-export const CompositionManager = createContext<CompositionManagerContext>({
-	compositions: [],
-	registerComposition: () => undefined,
-	unregisterComposition: () => undefined,
-	registerFolder: () => undefined,
-	unregisterFolder: () => undefined,
-	currentComposition: null,
-	setCurrentComposition: () => undefined,
-	setCurrentCompositionMetadata: () => undefined,
-	registerSequence: () => undefined,
-	unregisterSequence: () => undefined,
-	registerAsset: () => undefined,
-	unregisterAsset: () => undefined,
-	sequences: [],
-	assets: [],
-	folders: [],
-	currentCompositionMetadata: null,
-});
 
 export const compositionsRef = React.createRef<{
 	getCompositions: () => TCompMetadataWithCalcFunction<
