+++ packages/core/src/ResolveCompositionConfig.tsx
@@ -10,8 +10,7 @@ import {
 	useState,
 } from 'react';
 import type {AnyComposition} from './CompositionManager.js';
-// TODO: Cicular dependency
-import {CompositionManager} from './CompositionManager.js';
+import {CompositionManager} from './CompositionManagerContext.js';
 import {EditorPropsContext} from './EditorProps.js';
 import {resolveVideoConfig} from './resolve-video-config.js';
 import type {VideoConfig} from './video-config.js';
