+++ packages/core/src/test/audio.test.tsx
@@ -8,7 +8,7 @@ import {describe, expect, test} from 'vitest';
 import {z} from 'zod';
 import {Audio} from '../audio/index.js';
 import {CanUseRemotionHooksProvider} from '../CanUseRemotionHooks.js';
-import {CompositionManager} from '../CompositionManager.js';
+import {CompositionManager} from '../CompositionManagerContext.js';
 import {RemotionRoot} from '../RemotionRoot.js';
 import {ResolveCompositionConfig} from '../ResolveCompositionConfig.js';
