+++ packages/core/src/test/loop-validation.test.tsx
@@ -8,7 +8,7 @@ import {renderToString} from 'react-dom/server';
 import {describe, expect, test} from 'vitest';
 import {z} from 'zod';
 import {CanUseRemotionHooksProvider} from '../CanUseRemotionHooks.js';
-import {CompositionManager} from '../CompositionManager.js';
+import {CompositionManager} from '../CompositionManagerContext.js';
 import {Loop} from '../loop/index.js';
 import {RemotionRoot} from '../RemotionRoot.js';
 import {ResolveCompositionConfig} from '../ResolveCompositionConfig.js';
