+++ packages/core/src/test/video.test.tsx
@@ -7,7 +7,7 @@ import React, {useContext} from 'react';
 import {describe, expect, test} from 'vitest';
 import {z} from 'zod';
 import {CanUseRemotionHooksProvider} from '../CanUseRemotionHooks.js';
-import {CompositionManager} from '../CompositionManager.js';
+import {CompositionManager} from '../CompositionManagerContext.js';
 import {RemotionRoot} from '../RemotionRoot.js';
 import {ResolveCompositionConfig} from '../ResolveCompositionConfig.js';
 import {Video} from '../video/index.js';
